chapter3_4_aianimals/violation_detection/no_animal_violation/proxy/requirements.txt
absl-py==1.2.0; python_version >= "3.7"
anyio==3.6.1; python_full_version >= "3.6.2" and python_version >= "3.7"
astunparse==1.6.3; python_version >= "3.7"
cachetools==5.2.0; python_version >= "3.7" and python_version < "4.0" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7")
certifi==2022.6.15; python_version >= "3.7" and python_version < "4"
charset-normalizer==2.1.0; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0"
click==8.1.3; python_version >= "3.7"
colorama==0.4.5; python_version >= "3.7" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.5.0"
dependency-injector==4.40.0
flatbuffers==1.12; python_version >= "3.7"
gast==0.4.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
google-auth-oauthlib==0.4.6; python_version >= "3.7"
google-auth==2.10.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
google-pasta==0.2.0; python_version >= "3.7"
greenlet==1.1.2; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and python_full_version >= "3.5.0"
grpcio==1.48.0; python_version >= "3.7"
h11==0.12.0; python_version >= "3.7"
h5py==3.7.0; python_version >= "3.7"
httpcore==0.15.0; python_version >= "3.7"
httpx==0.23.0; python_version >= "3.7"
idna==3.3; python_version >= "3.7" and python_full_version >= "3.6.2" and python_version < "4"
keras-preprocessing==1.1.2; python_version >= "3.7"
keras==2.9.0; python_version >= "3.7"
libclang==14.0.6; python_version >= "3.7"
markdown==3.4.1; python_version >= "3.7"
markupsafe==2.1.1; python_version >= "3.7"
nptyping==1.4.4; python_version >= "3.5"
numpy==1.23.1; python_version >= "3.8"
oauthlib==3.2.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
opt-einsum==3.3.0; python_version >= "3.7"
packaging==21.3; python_version >= "3.7"
pika==1.3.0
pillow==9.2.0; python_version >= "3.7"
protobuf==3.19.6; python_version >= "3.7"
psycopg2-binary==2.9.3; python_version >= "3.6"
pyasn1-modules==0.2.8; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
pyasn1==0.4.8; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") or python_full_version >= "3.6.0" and python_version >= "3.7" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7")
pydantic==1.9.2; python_full_version >= "3.6.1"
pyparsing==3.0.9; python_full_version >= "3.6.8" and python_version >= "3.7"
requests-oauthlib==1.3.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
requests==2.28.1; python_version >= "3.7" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7")
rfc3986==1.5.0; python_version >= "3.7"
rsa==4.9; python_version >= "3.6" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7")
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
sniffio==1.2.0; python_full_version >= "3.6.2" and python_version >= "3.7"
sqlalchemy==1.4.40; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
tensorboard-data-server==0.6.1; python_version >= "3.7"
tensorboard-plugin-wit==1.8.1; python_version >= "3.7"
tensorboard==2.9.1; python_version >= "3.7"
tensorflow-estimator==2.9.0; python_version >= "3.7"
tensorflow-io-gcs-filesystem==0.26.0; python_version >= "3.7" and python_version < "3.11"
tensorflow-serving-api==2.9.1
tensorflow==2.9.1; python_version >= "3.7"
termcolor==1.1.0; python_version >= "3.7"
typing-extensions==4.3.0; python_version >= "3.7" and python_full_version >= "3.6.1"
typish==1.9.3; python_version >= "3.5"
urllib3==1.26.11; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7"
werkzeug==2.2.2; python_version >= "3.7"
wrapt==1.14.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
